+++ workers/unity/Packages/io.improbable.gdk.buildsystem/.codegen/Source/Generators/BuildSystem/BuildSystemAssemblyGenerator.cs
@@ -1,0 +1,22 @@
+namespace Improbable.Gdk.CodeGenerator
+{
+    public static class BuildSystemAssemblyGenerator
+    {
+        public static string Generate()
+        {
+            return @"{
+    ""name"": ""Improbable.Gdk.Generated.BuildSystem"",
+    ""references"": [
+        ""Improbable.Gdk.BuildSystem"",
+        ""Improbable.Gdk.Tools""
+    ],
+    ""includePlatforms"": [
+        ""Editor""
+    ],
+    ""excludePlatforms"": [],
+    ""allowUnsafeCode"": false
+}
+";
+        }
+    }
+}